Cleaned up table creation
Moved table creation for each repository to a static method. Then when the corresponding repo factory is created, it is able to ensure all the necessary tables are created. This helps with state management in that there should no longer be any missing tables discovered at run time, and prevents us from having to check for tables explicitly as part of processing each incoming request. Likewise, it allows us to eliminate quite a number of SqlMapper.AddTypeHandler calls.